requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
pillow
webcolors
python-multipart